Revert "[android] - configure loggin of HttpRequest, cleanup HttpRequest"
This reverts commit a059c74a54302f9604f3e3b28143be7645e6204f.
